+++ protocols/dcutr/src/handler.rs
@@ -23,7 +23,6 @@ use futures::future::{BoxFuture, FutureExt};
 use futures::stream::{FuturesUnordered, StreamExt};
 use libp2p_core::multiaddr::{Multiaddr, Protocol};
 use libp2p_core::{upgrade, ConnectedPoint, PeerId};
-use libp2p_swarm::protocols_handler::either::EitherHandler;
 use libp2p_swarm::protocols_handler::DummyProtocolsHandler;
 use libp2p_swarm::protocols_handler::{InboundUpgradeSend, OutboundUpgradeSend, SendWrapper};
 use libp2p_swarm::{
@@ -33,6 +32,7 @@ use libp2p_swarm::{
 use std::collections::VecDeque;
 use std::fmt;
 use std::task::{Context, Poll};
+use either::Either;
 
 pub enum In {
     Connect {
@@ -103,7 +103,7 @@ impl Prototype {
 }
 
 impl IntoProtocolsHandler for Prototype {
-    type Handler = EitherHandler<Handler, DummyProtocolsHandler>;
+    type Handler = Either<Handler, DummyProtocolsHandler>;
 
     fn into_handler(self, _remote_peer_id: &PeerId, endpoint: &ConnectedPoint) -> Self::Handler {
         if endpoint
@@ -111,13 +111,13 @@ impl IntoProtocolsHandler for Prototype {
             .iter()
             .any(|p| p == Protocol::P2pCircuit)
         {
-            EitherHandler::A(Handler {
+            Either::Left(Handler {
                 remote_addr: endpoint.get_remote_address().clone(),
                 queued_events: Default::default(),
                 inbound_connects: Default::default(),
             })
         } else {
-            EitherHandler::B(DummyProtocolsHandler::default())
+            Either::Right(DummyProtocolsHandler::default())
         }
     }