+++ muxers/yamux/CHANGELOG.md
@@ -1,5 +1,8 @@
 # 0.34.0 [unreleased]
 
+- Make default features of `libp2p-core` optional.
+  [PR 2181](https://github.com/libp2p/rust-libp2p/pull/2181)
+
 - Update dependencies.
 
 - Implement `From<io::Error> for YamuxError` instead of `Into`.